Some cleanup

+++ src/structures.h
@@ -355,7 +355,7 @@ struct struct_type_ope
   gint signe_type;   /* 0=neutre, 1=débit, 2=crédit */
   gint affiche_entree;    /* par ex pour virement, chèques ... */
   gint numerotation_auto;        /* pour les chèques */
-  gulong no_en_cours;
+  gdouble no_en_cours;
   gint no_compte;
 };
